Split offscreen live graph sync
