+++ web/js/export/render_graph_offscreen.mjs
@@ -1,6 +1,7 @@
 import { app } from "/scripts/app.js";
 import { computeGraphBBox } from "./bbox.mjs";
 import { applyBackgroundMode, getExportBackgroundFillColor } from "./background_modes.mjs";
+import { syncLiveGraphState } from "./live_graph_sync.mjs";
 import { collectNodeRects } from "../core/backends/legacy_bounds.mjs";
 import {
   drawDomWidgetOverlays,
@@ -14,13 +15,11 @@ import {
 } from "../core/backends/legacy_media_overlays.mjs";
 import {
   collectVideoElementsFromDom,
-  collectDomWidgetContainers,
   getCanvasRoot,
   getDomElementGraphRect,
   getNodeIdFromElement,
   isElementInGraphNode,
   collectDomMediaElements,
-  resolveNodeIdForGraphRect,
 } from "../core/overlays/dom_utils.mjs";
 
 const PREVIEW_MAX_PIXELS = 2048 * 2048;
@@ -320,407 +319,6 @@ function configureGraph(graph, workflowJson) {
   }
 }
 
-function copyNodeMedia(fromNode, toNode) {
-  if (!fromNode || !toNode) return false;
-  const mediaKeys = [
-    "imgs",
-    "img",
-    "image",
-    "preview",
-    "preview_image",
-    "previewImage",
-    "previewMediaType",
-    "canvas",
-    "previewCanvas",
-    "images",
-    "animatedImages",
-    "frames",
-    "frame",
-    "video_path",
-    "filepath",
-    "file",
-    "url",
-    "media",
-    "media_el",
-    "mediaEl",
-    "texture",
-    "tex",
-    "_texture",
-    "output_image",
-  ];
-  let copied = false;
-  for (const key of mediaKeys) {
-    if (fromNode[key] !== undefined && fromNode[key] !== null) {
-      // Never copy live HTML video elements into the export graph.
-      if (
-        key === "video" ||
-        key === "videos" ||
-        key === "videoEl" ||
-        key === "videoElement" ||
-        key === "media_el" ||
-        key === "mediaEl"
-      ) {
-        continue;
-      }
-      toNode[key] = fromNode[key];
-      copied = true;
-    }
-  }
-  return copied;
-}
-
-function syncLiveNodeMedia(exportGraph, liveGraph, debugLog) {
-  const liveNodes = liveGraph?._nodes || liveGraph?.nodes || [];
-  const exportNodes = exportGraph?._nodes || exportGraph?.nodes || [];
-  if (!liveNodes.length || !exportNodes.length) return;
-
-  const liveById = new Map();
-  for (const node of liveNodes) {
-    if (node && Number.isFinite(node.id)) {
-      liveById.set(node.id, node);
-    }
-  }
-
-  let copiedCount = 0;
-  for (const node of exportNodes) {
-    if (!node || !Number.isFinite(node.id)) continue;
-    const liveNode = liveById.get(node.id);
-    if (!liveNode) continue;
-    if (copyNodeMedia(liveNode, node)) {
-      copiedCount += 1;
-    }
-  }
-  debugLog?.("media.sync", { copiedCount });
-}
-
-function syncLiveNodeText(exportGraph, liveGraph) {
-  const liveNodes = liveGraph?._nodes || liveGraph?.nodes || [];
-  const exportNodes = exportGraph?._nodes || exportGraph?.nodes || [];
-  if (!liveNodes.length || !exportNodes.length) return;
-
-  const liveById = new Map();
-  for (const node of liveNodes) {
-    if (node && Number.isFinite(node.id)) {
-      liveById.set(node.id, node);
-    }
-  }
-
-  for (const node of exportNodes) {
-    if (!node || !Number.isFinite(node.id)) continue;
-    const liveNode = liveById.get(node.id);
-    if (!liveNode) continue;
-
-    if (liveNode.widgets_values !== undefined) {
-      node.widgets_values = liveNode.widgets_values;
-    }
-    if (liveNode.properties !== undefined) {
-      node.properties = liveNode.properties;
-    }
-    if (Number.isFinite(liveNode.widgets_start_y)) {
-      node.widgets_start_y = liveNode.widgets_start_y;
-    }
-    if (Array.isArray(node.widgets) && Array.isArray(liveNode.widgets)) {
-      const count = Math.min(node.widgets.length, liveNode.widgets.length);
-      const widgetsValues = liveNode.widgets_values;
-      const widgetsValuesKeys =
-        widgetsValues && typeof widgetsValues === "object" && !Array.isArray(widgetsValues)
-          ? Object.keys(widgetsValues)
-          : null;
-      for (let i = 0; i < count; i += 1) {
-        const exportWidget = node.widgets[i];
-        const liveWidget = liveNode.widgets[i];
-        if (!exportWidget || !liveWidget) continue;
-        const widgetName =
-          exportWidget.name ||
-          liveWidget.name ||
-          exportWidget?.options?.name ||
-          liveWidget?.options?.name;
-        let value = liveWidget.value;
-        if (value === undefined) {
-          if (widgetsValues && typeof widgetsValues === "object" && !Array.isArray(widgetsValues)) {
-            if (widgetName && widgetsValues[widgetName] !== undefined) {
-              value = widgetsValues[widgetName];
-            }
-          } else if (Array.isArray(widgetsValues) && widgetsValues[i] !== undefined) {
-            value = widgetsValues[i];
-          }
-        }
-        if (
-          value === undefined &&
-          widgetsValuesKeys &&
-          widgetsValuesKeys[i] !== undefined &&
-          widgetsValues[widgetsValuesKeys[i]] !== undefined
-        ) {
-          value = widgetsValues[widgetsValuesKeys[i]];
-        }
-        if (value === undefined && liveNode.properties && typeof liveNode.properties === "object") {
-          if (widgetName && liveNode.properties[widgetName] !== undefined) {
-            value = liveNode.properties[widgetName];
-          }
-        }
-
-        if (value !== undefined) {
-          if (typeof exportWidget.setValue === "function") {
-            try {
-              exportWidget.setValue(value);
-            } catch (_) {
-              // ignore setValue failures
-            }
-          } else {
-            try {
-              exportWidget.value = value;
-            } catch (_) {
-              // ignore read-only widget value
-            }
-            try {
-              exportWidget._value = value;
-            } catch (_) {
-              // ignore read-only _value
-            }
-            if (exportWidget.options && typeof exportWidget.options === "object") {
-              try {
-                exportWidget.options.value = value;
-              } catch (_) {
-                // ignore read-only options
-              }
-            }
-          }
-        }
-        if (Number.isFinite(liveWidget.y)) {
-          try {
-            exportWidget.y = liveWidget.y;
-          } catch (_) {
-            // ignore read-only widget y
-          }
-        }
-        if (Number.isFinite(liveWidget.height)) {
-          try {
-            exportWidget.height = liveWidget.height;
-          } catch (_) {
-            // ignore read-only widget height
-          }
-        }
-        if (Number.isFinite(liveWidget.aspectRatio)) {
-          try {
-            exportWidget.aspectRatio = liveWidget.aspectRatio;
-          } catch (_) {
-            // ignore read-only widget aspect ratio
-          }
-        }
-        if (Number.isFinite(liveWidget.computedHeight)) {
-          try {
-            exportWidget.computedHeight = liveWidget.computedHeight;
-          } catch (_) {
-            // ignore read-only widget computed height
-          }
-        }
-        if (liveWidget.parentEl && exportWidget.parentEl) {
-          try {
-            exportWidget.parentEl.hidden = Boolean(liveWidget.parentEl.hidden);
-          } catch (_) {
-            // ignore hidden sync failures
-          }
-        }
-      }
-    }
-    if (typeof node.computeSize === "function") {
-      try {
-        const nextSize = node.computeSize([node.size?.[0], node.size?.[1]]);
-        if (Array.isArray(nextSize) && nextSize.length >= 2) {
-          if (typeof node.setSize === "function") {
-            node.setSize([Number(nextSize[0]), Number(nextSize[1])]);
-          } else {
-            node.size = [Number(nextSize[0]), Number(nextSize[1])];
-          }
-        }
-      } catch (_) {
-        // ignore computeSize failures from custom nodes
-      }
-    }
-  }
-}
-
-function syncLiveNodeGeometry(exportGraph, liveGraph) {
-  const liveNodes = liveGraph?._nodes || liveGraph?.nodes || [];
-  const exportNodes = exportGraph?._nodes || exportGraph?.nodes || [];
-  if (!liveNodes.length || !exportNodes.length) return;
-
-  const liveById = new Map();
-  for (const node of liveNodes) {
-    if (node && Number.isFinite(node.id)) {
-      liveById.set(node.id, node);
-    }
-  }
-
-  const isValidPair = (pair) => Array.isArray(pair) && pair.length >= 2
-    && Number.isFinite(Number(pair[0]))
-    && Number.isFinite(Number(pair[1]));
-
-  for (const node of exportNodes) {
-    if (!node || !Number.isFinite(node.id)) continue;
-    const liveNode = liveById.get(node.id);
-    if (!liveNode) continue;
-
-    const livePos = liveNode.pos || liveNode._pos;
-    if (isValidPair(livePos)) {
-      node.pos = [Number(livePos[0]), Number(livePos[1])];
-    }
-
-    const liveSize = liveNode.size || liveNode._size;
-    if (isValidPair(liveSize)) {
-      node.size = [Number(liveSize[0]), Number(liveSize[1])];
-    }
-
-  }
-}
-
-function syncLiveDomWidgetHeights(exportGraph, uiCanvas) {
-  const exportNodes = exportGraph?._nodes || exportGraph?.nodes || [];
-  if (!exportNodes.length || !uiCanvas) return;
-
-  const exportById = new Map();
-  for (const node of exportNodes) {
-    if (node && Number.isFinite(node.id)) {
-      exportById.set(node.id, node);
-    }
-  }
-  if (!exportById.size) return;
-
-  const nodeRects = collectNodeRects(exportGraph, null);
-  const widgets = collectDomWidgetContainers(uiCanvas);
-  for (const widget of widgets) {
-    if (!(widget instanceof HTMLElement)) continue;
-    if (!widget.querySelector?.("video, img, canvas")) continue;
-
-    const rect = getDomElementGraphRect(widget, uiCanvas);
-    if (!rect || rect.h <= 0) continue;
-
-    const nodeId = resolveNodeIdForGraphRect(
-      nodeRects,
-      rect,
-      getNodeIdFromElement(widget)
-    );
-    if (!Number.isFinite(nodeId)) continue;
-
-    const exportNode = exportById.get(nodeId);
-    if (!exportNode) continue;
-
-    const nodePos = exportNode.pos || exportNode._pos;
-    const nodeSize = exportNode.size || exportNode._size;
-    if (!Array.isArray(nodePos) || !Array.isArray(nodeSize) || nodePos.length < 2 || nodeSize.length < 2) {
-      continue;
-    }
-
-    const requiredHeight = Math.ceil(rect.y + rect.h - Number(nodePos[1]));
-    if (Number.isFinite(requiredHeight) && requiredHeight > Number(nodeSize[1])) {
-      exportNode.size = [Number(nodeSize[0]), requiredHeight];
-    }
-  }
-
-  const videos = collectVideoElementsFromDom(uiCanvas);
-  for (const video of videos) {
-    if (!(video instanceof HTMLVideoElement)) continue;
-    const rect = getDomElementGraphRect(video, uiCanvas);
-    if (!rect || rect.h <= 0) continue;
-
-    const nodeId = resolveNodeIdForGraphRect(
-      nodeRects,
-      rect,
-      getNodeIdFromElement(video)
-    );
-    if (!Number.isFinite(nodeId)) continue;
-
-    const exportNode = exportById.get(nodeId);
-    if (!exportNode) continue;
-
-    const nodePos = exportNode.pos || exportNode._pos;
-    const nodeSize = exportNode.size || exportNode._size;
-    if (!Array.isArray(nodePos) || !Array.isArray(nodeSize) || nodePos.length < 2 || nodeSize.length < 2) {
-      continue;
-    }
-
-    const requiredHeight = Math.ceil(rect.y + rect.h - Number(nodePos[1]));
-    if (Number.isFinite(requiredHeight) && requiredHeight > Number(nodeSize[1])) {
-      exportNode.size = [Number(nodeSize[0]), requiredHeight];
-    }
-  }
-}
-
-function syncLiveGroups(exportGraph, liveGraph) {
-  const exportGroups = exportGraph?._groups || exportGraph?.groups || [];
-  const liveGroups = liveGraph?._groups || liveGraph?.groups || [];
-  if (!exportGroups.length || !liveGroups.length) return;
-
-  const normalizePos = (pos) => {
-    if (Array.isArray(pos) && pos.length >= 2) return [pos[0], pos[1]];
-    return null;
-  };
-  const normalizeSize = (size) => {
-    if (Array.isArray(size) && size.length >= 2) return [size[0], size[1]];
-    return null;
-  };
-  const distanceSq = (a, b) => {
-    if (!a || !b) return Number.POSITIVE_INFINITY;
-    const dx = a[0] - b[0];
-    const dy = a[1] - b[1];
-    return dx * dx + dy * dy;
-  };
-
-  const liveById = new Map();
-  for (const group of liveGroups) {
-    if (!group) continue;
-    if (group.id !== undefined && group.id !== null) {
-      liveById.set(group.id, group);
-    }
-  }
-
-  for (const exportGroup of exportGroups) {
-    if (!exportGroup) continue;
-    let liveGroup = null;
-
-    if (exportGroup.id !== undefined && exportGroup.id !== null) {
-      liveGroup = liveById.get(exportGroup.id) || null;
-    }
-
-    if (!liveGroup && exportGroup.title) {
-      const sameTitle = liveGroups.filter((g) => g?.title === exportGroup.title);
-      if (sameTitle.length === 1) {
-        liveGroup = sameTitle[0];
-      } else if (sameTitle.length > 1) {
-        const exportPos = normalizePos(exportGroup.pos || exportGroup._pos);
-        let best = null;
-        let bestDist = Number.POSITIVE_INFINITY;
-        for (const candidate of sameTitle) {
-          const candPos = normalizePos(candidate.pos || candidate._pos);
-          const dist = distanceSq(exportPos, candPos);
-          if (dist < bestDist) {
-            bestDist = dist;
-            best = candidate;
-          }
-        }
-        liveGroup = best;
-      }
-    }
-
-    if (!liveGroup) {
-      // fallback to same index if nothing else matches
-      const idx = exportGroups.indexOf(exportGroup);
-      liveGroup = liveGroups[idx] || null;
-    }
-
-    if (!liveGroup) continue;
-
-    const livePos = normalizePos(liveGroup.pos || liveGroup._pos);
-    const liveSize = normalizeSize(liveGroup.size || liveGroup._size);
-    if (livePos) {
-      exportGroup.pos = [...livePos];
-    }
-    if (liveSize) {
-      exportGroup.size = [...liveSize];
-    }
-  }
-}
-
 function configureTransform(offscreen, bbox, padding) {
   if (!offscreen || !offscreen.ds) {
     return;
@@ -922,11 +520,7 @@ async function prepareGraph(workflowJson, debugLog) {
   }
   const graph = new LGraphRef();
   configureGraph(graph, workflowJson);
-  syncLiveNodeGeometry(graph, app?.graph);
-  syncLiveNodeMedia(graph, app?.graph, debugLog);
-  syncLiveNodeText(graph, app?.graph);
-  syncLiveDomWidgetHeights(graph, app?.canvas);
-  syncLiveGroups(graph, app?.graph);
+  syncLiveGraphState(graph, app?.graph, app?.canvas, debugLog);
   return { graph, LGraphCanvasRef };
 }
